Simplify promptframe library, remove complex files, add minimal example
